Building scripts for vagrant box

+++ tools/clean_build.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
-shopt -s extglob
-cd ../build
-rm -rfv !("." || "..")
-shopt -u extglob
 
-cmake ../ -DCMAKE_BUILD_TYPE=debug
-cmake --build .
+shopt -s extglob
+if [ "${C}" ] && [ "${CXX}" ];
+then
+    cd ../build
+    rm -rfv !("." || "..")
+    cmake .. -DCMAKE_BUILD_TYPE=Debug
+    cmake --build . --parallel 2 --config Debug
+else
+    echo "Kit not selected !"
+fi
+shopt -u extglob
